This makes fields for time that have readWrite access writable. (#199)
* This makes fields for time that have readWrite access writable.

Some examples of those are "BSH.Common.Option.Duration" and "Cooking.Oven.Setting.Cavity.001.AlarmClock" on ovens.

The first one is the one that determines the time after which the oven must be turned on.
The second one is similar but it doesn't turn off the oven, it just makes an alarm sound but the oven stays on.

* Fix typo

* Remove redundant check